+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 
 - **[FIR]**: Disallow `_` assisted context parameter names in top-level function injection.
 - **[FIR/IR]**: When generating class and provider factories now, the compiler dedupes non-assisted, non-optional injections of the same type key (i.e. type ± qualifier). This shrinks generated code size in (uncommon) scenarios where you inject the same type multiple types.
+- **[Gradle]**: Allow `DiagnosticSeverity` metro extension properties to be configurable as `metro.*` gradle properties of the same name.
 
 ### Fixes